requirements.txt
dill==0.3.8
dnspython==2.6.1
fastapi==0.111.0
httptools==0.6.1
imblearn==0.0
mypy-boto3-s3==1.34.138
pip-chill==1.0.3
pymongo==4.8.0
python-dotenv==1.0.1
types-s3transfer==0.10.1
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
wincertstore==0.2.1
xgboost==2.1.0
pandas==2.2.2
PyYAML==6.0.1
numpy==2.0.0
scikit-learn==1.5.1
apache-airflow==2.9.2
-e .
